+++ src/Foody/Foody.WebApi/Controllers/v1/ProductController.cs
@@ -1,6 +1,4 @@
-﻿using AutoMapper;
-using Foody.Data.Services;
-using Foody.Data.Interfaces;
+﻿using Foody.Data.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -9,10 +7,10 @@ namespace Foody.WebApi.Controllers.v1
 {
     public sealed class ProductController : BaseController
     {
-        public ProductController(IUnitofWork unitofWork, IMapper mapper, ICacheService cacheService)
+        public ProductController(IUnitOfWork unitofWork, IMapper mapper, ICacheService cacheService)
             : base(unitofWork, mapper, cacheService)
         {
-            this._cached = "products";
+            _cached = "products";
         }
 
         // GET: v1/Product
@@ -21,12 +19,12 @@ namespace Foody.WebApi.Controllers.v1
         [ProducesResponseType(200, Type = typeof(Pagination<ProductDto>))]
         public async Task<IActionResult> Get(string? search)
         {
-            bool cacheEnable = string.IsNullOrEmpty(search);
+            bool cacheEnable = string.IsNullOrWhiteSpace(search);
 
             var cacheData = cacheEnable ? await GetCache<IEnumerable<ProductDto>>(_cached) : null;
             if (cacheData is null)
             {
-                var products = await _unitofWork.Products.All(search);
+                var products = await _unitofWork.Products.AllAsync(search!);
                 var _dto = _mapper.Map<List<ProductDto>>(products);
                 cacheData = _dto;
                 if (cacheEnable)
@@ -37,26 +35,26 @@ namespace Foody.WebApi.Controllers.v1
         }
 
         // GET v1/Product/ByCategory
-        [HttpGet]
-        [Route("ByCategory")]
-        [ProducesResponseType(200 , Type = typeof(Result<Dictionary<string, List<ProdCategoryDto>>>))]
-        public async Task<IActionResult> ByCategory()
-        {
-            var result = new Result<dynamic>();
+        //[HttpGet]
+        //[Route("ByCategory")]
+        //[ProducesResponseType(200 , Type = typeof(Result<Dictionary<string, List<ProdCategoryDto>>>))]
+        //public async Task<IActionResult> ByCategory()
+        //{
+        //    var result = new Result<dynamic>();
 
-            var dict = await _unitofWork.Products.ProducstByCategory();
-            if (dict is null)
-            {
-                result.Error = AddError(500,
-                    ErrorsMessage.Generic.NullSet,
-                    ErrorsMessage.Generic.UnknownError);
+        //    var dict = await _unitofWork.Products.ProducstByCategory();
+        //    if (dict is null)
+        //    {
+        //        result.Error = AddError(500,
+        //            ErrorsMessage.Generic.NullSet,
+        //            ErrorsMessage.Generic.UnknownError);
 
-                return StatusCode(500, result);
-            }
+        //        return StatusCode(500, result);
+        //    }
 
-            result.Content = dict;
-            return Ok(result);
-        }
+        //    result.Content = dict;
+        //    return Ok(result);
+        //}
 
         // GET v1/Product/5
         [HttpGet("{id}")]
@@ -69,7 +67,7 @@ namespace Foody.WebApi.Controllers.v1
             var cacheData = await GetCache<ProductDetailDto>(key);
             if (cacheData is null)
             {
-                var product = await _unitofWork.Products.Get(id);
+                var product = await _unitofWork.Products.GetAsync(id);
 
                 if (product is null)
                 {
@@ -86,7 +84,7 @@ namespace Foody.WebApi.Controllers.v1
             }
 
             result.Content = cacheData;
-            return Ok(result); 
+            return Ok(result);
         }
 
         // POST v1/Product
@@ -96,7 +94,7 @@ namespace Foody.WebApi.Controllers.v1
         {
             var result = new Result<ProductDto>();
 
-            if (!await _unitofWork.Products.Exists(productModDto.Name))
+            if (!await _unitofWork.Products.ExistsAsync(productModDto.Name))
             {
                 var product = _mapper.Map<Product>(productModDto);
                 await Upload(product, productModDto.ImageUpload);
@@ -104,14 +102,14 @@ namespace Foody.WebApi.Controllers.v1
                 var invalid = ValidateModel(product, productModDto, result);
                 if (invalid is not null) return invalid;
 
-                await _unitofWork.Products.Add(product);
-                await _unitofWork.CompleteAsync();
+                await _unitofWork.Products.AddAsync(product);
+                await _unitofWork.CommitAsync();
 
                 var _dto = _mapper.Map<ProductDto>(product);
                 //await SetCache($"{_dto.Id}", _dto, _cached);
 
                 result.Content = _dto;
-                return CreatedAtAction(nameof(Get), new { id = product.Id }, result); 
+                return CreatedAtAction(nameof(Get), new { id = product.Id }, result);
             }
 
             result.Error = AddError(409,
@@ -128,7 +126,7 @@ namespace Foody.WebApi.Controllers.v1
         {
             var result = new Result<dynamic>();
 
-            var product = await _unitofWork.Products.Get(id);
+            var product = await _unitofWork.Products.GetAsync(id);
             if (product is null || product.Id != id)
             {
                 result.Error = AddError(404,
@@ -138,7 +136,7 @@ namespace Foody.WebApi.Controllers.v1
                 return NotFound(result);
             }
 
-            if (await _unitofWork.Categories.Exists(productModDto.Name) && productModDto.Name != product.Name)
+            if (await _unitofWork.Categories.ExistsAsync(productModDto.Name) && productModDto.Name != product.Name)
             {
                 result.Error = AddError(409,
                    ErrorsMessage.Generic.ValidationError,
@@ -155,8 +153,8 @@ namespace Foody.WebApi.Controllers.v1
                 var invalid = ValidateModel(product, productModDto, result);
                 if (invalid is not null) return invalid;
 
-                await _unitofWork.Products.Update(product);
-                await _unitofWork.CompleteAsync();
+                await _unitofWork.Products.UpdateAsync(product);
+                await _unitofWork.CommitAsync();
                 //await SetCache($"{}", null, _cached);
             }
             catch (DbUpdateConcurrencyException) when (!_unitofWork.Products.Exists(id))
@@ -176,22 +174,20 @@ namespace Foody.WebApi.Controllers.v1
         [ProducesResponseType(204)]
         public async Task<IActionResult> Delete(int id)
         {
-            try
+            bool deleted = await _unitofWork.Products.DeleteAsync(id);
+            if (!deleted)
             {
-                await _unitofWork.Products.Delete(id);
-                await _unitofWork.CompleteAsync();
-                await DeleteCache($"{id}");
-                await DeleteCache(_cached);
-            }
-            catch (NullReferenceException)
-            {
-                Result<dynamic> result = new();
-                result.Error = AddError(404,
+                return NotFound(new Result<object>
+                {
+                    Error = AddError(404,
                     ErrorsMessage.Generic.NotFound,
-                    ErrorsMessage.Product.NotExist);
-
-                return NotFound(result);
+                    ErrorsMessage.Category.NotExist)
+                });
             }
+
+            await _unitofWork.CommitAsync();
+            await DeleteCache($"{_cached}-{id}");
+            await DeleteCache(_cached);
 
             return NoContent();
         }
